Reorganize directory to avoid abundant nesting
