+++ backend/src/index.ts
@@ -1,14 +1,14 @@
 import cors from "@fastify/cors";
 import jwt from "@fastify/jwt";
 import fastify, { FastifyInstance } from "fastify";
-import dbPlugin from "./modules/db/db.plugin";
-import wsPlugin from "./modules/ws/ws.plugin";
-import authPlugin from "./modules/auth/auth.plugin";
-import userPlugin from "./modules/user/user.plugin";
-import gamePlugin from "./modules/game/game.plugin";
-import tournamentPlugin from "./modules/tournament/tournament.plugin";
-import { seed } from "./modules/db/db.seed";
-import { config } from "./config";
+import dbPlugin from "./db/db.plugin";
+import wsPlugin from "./ws/ws.plugin";
+import authPlugin from "./auth/auth.plugin";
+import userPlugin from "./user/user.plugin";
+import gamePlugin from "./game/game.plugin";
+import tournamentPlugin from "./tournament/tournament.plugin";
+import { seed } from "./db/db.seed";
+import { config } from "./utils/config";
 
 let app: FastifyInstance | null = null;
 
@@ -20,7 +20,7 @@ try {
 
     // CORS
     // TODO: When in production what should the origin be? SITES? are env vars loaded?
-    await app.register(cors, { origin: config.isDev ? "*" : "placeholder" });
+    await app.register(cors, { origin: config.isProd ? "placeholder" : "*" });
 
     // TODO: Add more options here for JWT, see: https://github.com/fastify/fastify-jwt
     await app.register(jwt, { secret: config.jwtSecret }); // Register jwt plugin
@@ -32,8 +32,8 @@ try {
     await app.register(gamePlugin); // Register game plugin
     await app.register(tournamentPlugin); // Register tournament plugin
 
-    // Seed database in dev mode
-    if (config.isDev) await seed(app);
+    // Seed database if not in production
+    if (!config.isProd) await seed(app);
 
     // Start server
     await app.listen({ port: config.port, host: "0.0.0.0" });
