--- backend/test/tsconfig.json
@@ -1,7 +1,0 @@
-{
-    "extends": "../tsconfig.json",
-    "compilerOptions": {
-        "allowImportingTsExtensions": true
-    },
-    "include": ["**/*.test.ts"]
-}
