+++ backend/tsconfig.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
-        "target": "es2020",
-        "module": "es2022",
+        "target": "ES2022",
+        "module": "ESNext",
         "moduleResolution": "Bundler",
         "strict": true,
         "esModuleInterop": true,
